[RENAME] MemoCardCollectionViewCell <-> MemoCollectionViewCell 이름 변경 #23
